refactor: acl helpers and global tables
