+++ src/server/acl/acl_family.cc
@@ -3,8 +3,6 @@
 
 #include "server/acl/acl_family.h"
 
-#include <glog/logging.h>
-
 #include <algorithm>
 #include <cctype>
 #include <chrono>
@@ -20,10 +18,11 @@
 
 #include "absl/container/flat_hash_set.h"
 #include "absl/flags/commandlineflag.h"
+#include "absl/strings/escaping.h"
 #include "absl/strings/match.h"
 #include "absl/strings/numbers.h"
 #include "absl/strings/str_cat.h"
-#include "absl/types/span.h"
+#include "absl/strings/str_split.h"
 #include "base/flags.h"
 #include "base/logging.h"
 #include "core/overloaded.h"
@@ -31,10 +30,8 @@
 #include "facade/facade_types.h"
 #include "io/file.h"
 #include "io/file_util.h"
-#include "io/io.h"
 #include "server/acl/acl_commands_def.h"
 #include "server/acl/acl_log.h"
-#include "server/acl/helpers.h"
 #include "server/acl/validator.h"
 #include "server/command_registry.h"
 #include "server/common.h"
@@ -116,9 +113,10 @@ void AclFamily::SetUser(CmdArgList args, ConnectionContext* cntx) {
     if (!exists) {
       User::UpdateRequest default_req;
       default_req.updates = {User::UpdateRequest::CategoryValueType{User::Sign::MINUS, acl::ALL}};
-      user.Update(std::move(default_req));
+      user.Update(std::move(default_req), CategoryToIdx(), reverse_cat_table_,
+                  CategoryToCommandsIndex());
     }
-    user.Update(std::move(req));
+    user.Update(std::move(req), CategoryToIdx(), reverse_cat_table_, CategoryToCommandsIndex());
     if (exists) {
       StreamUpdatesToAllProactorConnections(std::string(username), user.AclCommands(), user.Keys());
     }
@@ -300,13 +298,16 @@ GenericError AclFamily::LoadToRegistryFromFile(std::string_view full_path,
     User::UpdateRequest default_req;
     default_req.updates = {User::UpdateRequest::CategoryValueType{User::Sign::MINUS, acl::ALL}};
     auto& user = registry[usernames[i]];
-    user.Update(std::move(default_req));
-    user.Update(std::move(requests[i]));
+    user.Update(std::move(default_req), CategoryToIdx(), reverse_cat_table_,
+                CategoryToCommandsIndex());
+    user.Update(std::move(requests[i]), CategoryToIdx(), reverse_cat_table_,
+                CategoryToCommandsIndex());
   }
 
   if (!registry.contains("default")) {
     auto& user = registry["default"];
-    user.Update(registry_->DefaultUserUpdateRequest());
+    user.Update(registry_->DefaultUserUpdateRequest(), CategoryToIdx(), reverse_cat_table_,
+                CategoryToCommandsIndex());
   }
 
   return {};
@@ -440,13 +441,13 @@ void AclFamily::Cat(CmdArgList args, ConnectionContext* cntx) {
   if (args.size() == 1) {
     ToUpper(&args[0]);
     std::string_view category = facade::ToSV(args[0]);
-    if (!CATEGORY_INDEX_TABLE.contains(category)) {
+    if (!cat_table_.contains(category)) {
       auto error = absl::StrCat("Unkown category: ", category);
       cntx->SendError(error);
       return;
     }
 
-    const uint32_t cid_mask = CATEGORY_INDEX_TABLE.find(category)->second;
+    const uint32_t cid_mask = cat_table_.find(category)->second;
     std::vector<std::string_view> results;
     // TODO replace this with indexer
     auto cb = [cid_mask, &results](auto name, auto& cid) {
@@ -466,7 +467,7 @@ void AclFamily::Cat(CmdArgList args, ConnectionContext* cntx) {
   }
 
   size_t total_categories = 0;
-  for (auto& elem : REVERSE_CATEGORY_INDEX_TABLE) {
+  for (auto& elem : reverse_cat_table_) {
     if (elem != "_RESERVED") {
       ++total_categories;
     }
@@ -474,7 +475,7 @@ void AclFamily::Cat(CmdArgList args, ConnectionContext* cntx) {
 
   auto* rb = static_cast<facade::RedisReplyBuilder*>(cntx->reply_builder());
   rb->StartArray(total_categories);
-  for (auto& elem : REVERSE_CATEGORY_INDEX_TABLE) {
+  for (auto& elem : reverse_cat_table_) {
     if (elem != "_RESERVED") {
       rb->SendSimpleString(elem);
     }
@@ -638,6 +639,9 @@ void AclFamily::Register(dfly::CommandRegistry* registry) {
   *registry << CI{"ACL GENPASS", CO::NOSCRIPT | CO::LOADING, -1, 0, 0, acl::kGenPass}.HFUNC(
       GenPass);
   cmd_registry_ = registry;
+
+  // build indexers
+  BuildIndexers(cmd_registry_->GetFamilies());
 }
 
 #undef HFUNC
@@ -655,7 +659,7 @@ void AclFamily::Init(facade::Listener* main_listener, UserRegistry* registry) {
   if (!acl_file.empty() && Load()) {
     return;
   }
-  registry_->Init();
+  registry_->Init(&CategoryToIdx(), &reverse_cat_table_, &CategoryToCommandsIndex());
   config_registry.RegisterMutable("aclfile");
   config_registry.RegisterMutable("acllog_max_len", [this](const absl::CommandLineFlag& flag) {
     auto res = flag.TryGet<size_t>();
@@ -666,6 +670,363 @@ void AclFamily::Init(facade::Listener* main_listener, UserRegistry* registry) {
     }
     return res.has_value();
   });
+}
+
+std::string AclFamily::AclCatToString(uint32_t acl_category, User::Sign sign) const {
+  std::string res = sign == User::Sign::PLUS ? "+@" : "-@";
+  if (acl_category == acl::ALL) {
+    absl::StrAppend(&res, "all");
+    return res;
+  }
+
+  const auto& index = CategoryToIdx().at(acl_category);
+  absl::StrAppend(&res, absl::AsciiStrToLower(reverse_cat_table_[index]));
+  return res;
+}
+
+std::string AclFamily::AclCommandToString(size_t family, uint64_t mask, User::Sign sign) const {
+  // This is constant but can be optimized with an indexer
+  const auto& rev_index = CommandsRevIndexer();
+  std::string res;
+  std::string prefix = (sign == User::Sign::PLUS) ? "+" : "-";
+  if (mask == ALL_COMMANDS) {
+    for (const auto& cmd : rev_index[family]) {
+      absl::StrAppend(&res, prefix, absl::AsciiStrToLower(cmd), " ");
+    }
+    res.pop_back();
+    return res;
+  }
+
+  size_t pos = 0;
+  while (mask != 0) {
+    ++pos;
+    mask = mask >> 1;
+  }
+  --pos;
+  absl::StrAppend(&res, prefix, absl::AsciiStrToLower(rev_index[family][pos]));
+  return res;
+}
+
+AclFamily::MergeResult AclFamily::MergeTables(const User::CategoryChanges& categories,
+                                              const User::CommandChanges& commands) const {
+  MergeResult result;
+  for (auto [cat, meta] : categories) {
+    result.push_back(CategoryAndMetadata{cat, meta});
+  }
+
+  for (auto [cmd, meta] : commands) {
+    result.push_back(CommandAndMetadata{cmd, meta});
+  }
+
+  std::sort(result.begin(), result.end(), [](const auto& l, const auto& r) {
+    auto fetch = [](const auto& l) { return l.metadata.seq_no; };
+    return std::visit(fetch, l) < std::visit(fetch, r);
+  });
+
+  return result;
+}
+
+std::string AclFamily::AclCatAndCommandToString(const User::CategoryChanges& cat,
+                                                const User::CommandChanges& cmds) const {
+  std::string result;
+
+  auto tables = MergeTables(cat, cmds);
+
+  auto cat_visitor = [&result, this](const CategoryAndMetadata& val) {
+    const auto& [change, meta] = val;
+    absl::StrAppend(&result, AclCatToString(change, meta.sign), " ");
+  };
+
+  auto cmd_visitor = [&result, this](const CommandAndMetadata& val) {
+    const auto& [change, meta] = val;
+    const auto [family, bit_index] = change;
+    absl::StrAppend(&result, AclCommandToString(family, bit_index, meta.sign), " ");
+  };
+
+  Overloaded visitor{cat_visitor, cmd_visitor};
+
+  for (auto change : tables) {
+    std::visit(visitor, change);
+  }
+
+  if (!result.empty()) {
+    result.pop_back();
+  }
+
+  return result;
+}
+
+std::string AclFamily::PrettyPrintSha(std::string_view pass, bool all) const {
+  if (all) {
+    return absl::BytesToHexString(pass);
+  }
+  return absl::BytesToHexString(pass.substr(0, 15)).substr(0, 15);
+};
+
+std::optional<AclFamily::ParseKeyResult> AclFamily::MaybeParseAclKey(
+    std::string_view command) const {
+  if (absl::EqualsIgnoreCase(command, "ALLKEYS") || command == "~*") {
+    return ParseKeyResult{"", {}, true};
+  }
+
+  if (absl::EqualsIgnoreCase(command, "RESETKEYS")) {
+    return ParseKeyResult{"", {}, false, true};
+  }
+
+  auto op = KeyOp::READ_WRITE;
+
+  if (absl::StartsWith(command, "%RW")) {
+    command = command.substr(3);
+  } else if (absl::StartsWith(command, "%R")) {
+    op = KeyOp::READ;
+    command = command.substr(2);
+  } else if (absl::StartsWith(command, "%W")) {
+    op = KeyOp::WRITE;
+    command = command.substr(2);
+  }
+
+  if (!absl::StartsWith(command, "~")) {
+    return {};
+  }
+
+  auto key = command.substr(1);
+  if (key.empty()) {
+    return {};
+  }
+  return ParseKeyResult{std::string(key), op};
+}
+
+std::optional<User::UpdatePass> AclFamily::MaybeParsePassword(std::string_view command,
+                                                              bool hashed) const {
+  using UpPass = User::UpdatePass;
+  if (command == "nopass") {
+    return UpPass{"", false, true};
+  }
+
+  if (command == "resetpass") {
+    return UpPass{"", false, false, true};
+  }
+
+  if (command[0] == '>' || (hashed && command[0] == '#')) {
+    return UpPass{std::string(command.substr(1))};
+  }
+
+  if (command[0] == '<') {
+    return UpPass{std::string(command.substr(1)), true};
+  }
+
+  return {};
+}
+
+std::optional<bool> AclFamily::MaybeParseStatus(std::string_view command) const {
+  if (command == "ON") {
+    return true;
+  }
+  if (command == "OFF") {
+    return false;
+  }
+  return {};
+}
+
+using OptCat = std::optional<uint32_t>;
+
+// bool == true if +
+// bool == false if -
+std::pair<OptCat, bool> AclFamily::MaybeParseAclCategory(std::string_view command) const {
+  if (absl::StartsWith(command, "+@")) {
+    auto res = cat_table_.find(command.substr(2));
+    if (res == cat_table_.end()) {
+      return {};
+    }
+    return {res->second, true};
+  }
+
+  if (absl::StartsWith(command, "-@")) {
+    auto res = cat_table_.find(command.substr(2));
+    if (res == cat_table_.end()) {
+      return {};
+    }
+    return {res->second, false};
+  }
+
+  return {};
+}
+
+std::pair<AclFamily::OptCommand, bool> AclFamily::MaybeParseAclCommand(
+    std::string_view command, const CommandRegistry& registry) const {
+  if (absl::StartsWith(command, "+")) {
+    auto res = registry.Find(command.substr(1));
+    if (!res) {
+      return {};
+    }
+    std::pair<size_t, uint64_t> cmd{res->GetFamily(), res->GetBitIndex()};
+    return {cmd, true};
+  }
+
+  if (absl::StartsWith(command, "-")) {
+    auto res = registry.Find(command.substr(1));
+    if (!res) {
+      return {};
+    }
+    std::pair<size_t, uint64_t> cmd{res->GetFamily(), res->GetBitIndex()};
+    return {cmd, false};
+  }
+
+  return {};
+}
+
+AclFamily::MaterializedContents AclFamily::MaterializeFileContents(
+    std::vector<std::string>* usernames, std::string_view file_contents) const {
+  // This is fine, a very large file will top at 1-2 mb. And that's for 5000+ users with 400
+  // characters per line
+  std::vector<std::string_view> commands = absl::StrSplit(file_contents, "\n");
+  std::vector<std::vector<std::string_view>> materialized;
+  materialized.reserve(commands.size());
+  usernames->reserve(commands.size());
+  for (auto& command : commands) {
+    if (command.empty())
+      continue;
+    std::vector<std::string_view> cmds = absl::StrSplit(command, ' ');
+    if (!absl::EqualsIgnoreCase(cmds[0], "USER") || cmds.size() < 4) {
+      return {};
+    }
+
+    usernames->push_back(std::string(cmds[1]));
+    cmds.erase(cmds.begin(), cmds.begin() + 2);
+    materialized.push_back(cmds);
+  }
+  return materialized;
+}
+
+using facade::ErrorReply;
+
+std::variant<User::UpdateRequest, ErrorReply> AclFamily::ParseAclSetUser(
+    facade::ArgRange args, const CommandRegistry& registry, bool hashed, bool has_all_keys) const {
+  User::UpdateRequest req;
+
+  for (std::string_view arg : args) {
+    if (auto pass = MaybeParsePassword(facade::ToSV(arg), hashed); pass) {
+      req.passwords.push_back(std::move(*pass));
+
+      if (hashed && absl::StartsWith(facade::ToSV(arg), "#")) {
+        req.passwords.back().is_hashed = true;
+      }
+      continue;
+    }
+
+    if (auto res = MaybeParseAclKey(facade::ToSV(arg)); res) {
+      auto& [glob, op, all_keys, reset_keys] = *res;
+      if ((has_all_keys && !all_keys && !reset_keys) ||
+          (req.allow_all_keys && !all_keys && !reset_keys)) {
+        return ErrorReply(
+            "Error in ACL SETUSER modifier '~tmp': Adding a pattern after the * pattern (or the "
+            "'allkeys' flag) is not valid and does not have any effect. Try 'resetkeys' to start "
+            "with an empty list of patterns");
+      }
+
+      req.allow_all_keys = all_keys;
+      req.reset_all_keys = reset_keys;
+      if (reset_keys) {
+        has_all_keys = false;
+      }
+      req.keys.push_back({std::move(glob), op, all_keys, reset_keys});
+      continue;
+    }
+
+    std::string command = absl::AsciiStrToUpper(arg);
+
+    if (auto status = MaybeParseStatus(command); status) {
+      if (req.is_active) {
+        return ErrorReply("Multiple ON/OFF are not allowed");
+      }
+      req.is_active = *status;
+      continue;
+    }
+
+    auto [cat, add] = MaybeParseAclCategory(command);
+    if (cat) {
+      using Sign = User::Sign;
+      using Val = std::pair<Sign, uint32_t>;
+      auto val = add ? Val{Sign::PLUS, *cat} : Val{Sign::MINUS, *cat};
+      req.updates.push_back(val);
+      continue;
+    }
+
+    auto [cmd, sign] = MaybeParseAclCommand(command, registry);
+    if (!cmd) {
+      return ErrorReply(absl::StrCat("Unrecognized parameter ", command));
+    }
+
+    using Sign = User::Sign;
+    using Val = User::UpdateRequest::CommandsValueType;
+    auto [index, bit] = *cmd;
+    auto val = sign ? Val{Sign::PLUS, index, bit} : Val{Sign::MINUS, index, bit};
+    req.updates.push_back(val);
+  }
+
+  return req;
+}
+
+using facade::CmdArgList;
+
+std::string AclFamily::AclKeysToString(const AclKeys& keys) const {
+  if (keys.all_keys) {
+    return "~*";
+  }
+  std::string result;
+  for (auto& [pattern, op] : keys.key_globs) {
+    if (op == KeyOp::READ_WRITE) {
+      absl::StrAppend(&result, "~", pattern, " ");
+      continue;
+    }
+    std::string op_str = (op == KeyOp::READ) ? "R" : "W";
+    absl::StrAppend(&result, "%", op_str, "~", pattern, " ");
+  }
+
+  if (!result.empty()) {
+    result.pop_back();
+  }
+  return result;
+}
+
+std::string AclFamily::PasswordsToString(const absl::flat_hash_set<std::string>& passwords,
+                                         bool nopass, bool full_sha) const {
+  if (nopass) {
+    return "nopass ";
+  }
+  std::string result;
+  for (const auto& pass : passwords) {
+    absl::StrAppend(&result, "#", PrettyPrintSha(pass, full_sha), " ");
+  }
+
+  return result;
+}
+
+void AclFamily::BuildIndexers(RevCommandsIndexStore families) {
+  acl::NumberOfFamilies(families.size());
+  CommandsRevIndexer(std::move(families));
+  CategoryToCommandsIndexStore index;
+  cmd_registry_->Traverse([&](std::string_view name, auto& cid) {
+    auto cat = cid.acl_categories();
+    for (size_t i = 0; i < 32; ++i) {
+      if (cat & (1 << i)) {
+        std::string_view cat_name = reverse_cat_table_[i];
+        if (index[cat_name].empty()) {
+          index[cat_name].resize(CommandsRevIndexer().size());
+        }
+        auto family = cid.GetFamily();
+        auto bit_index = cid.GetBitIndex();
+        index[cat_name][family] |= bit_index;
+      }
+    }
+  });
+
+  CategoryToCommandsIndex(std::move(index));
+  CategoryToIdxStore idx_store;
+  for (size_t i = 0; i < 32; ++i) {
+    idx_store[1 << i] = i;
+  }
+  CategoryToIdx(std::move(idx_store));
 }
 
 }  // namespace dfly::acl
