+++ src/server/acl/acl_family.h
@@ -13,6 +13,7 @@
 #include "facade/dragonfly_listener.h"
 #include "facade/facade_types.h"
 #include "helio/util/proactor_pool.h"
+#include "server/acl/acl_commands_def.h"
 #include "server/acl/user_registry.h"
 #include "server/command_registry.h"
 #include "server/common.h"
@@ -62,12 +63,154 @@ class AclFamily final {
   // Helper function that loads the acl state of an acl file into the user registry
   GenericError LoadToRegistryFromFile(std::string_view full_path, ConnectionContext* init);
 
+  // Serializes the whole registry into a string
   std::string RegistryToString() const;
+
+  std::string AclCatToString(uint32_t acl_category, User::Sign sign) const;
+
+  std::string AclCommandToString(size_t family, uint64_t mask, User::Sign sign) const;
+
+  // Serializes category and command to string
+  std::string AclCatAndCommandToString(const User::CategoryChanges& cat,
+                                       const User::CommandChanges& cmds) const;
+
+  std::string PrettyPrintSha(std::string_view pass, bool all = false) const;
+
+  // When hashed is true, we allow passwords that start with both # and >
+  std::optional<User::UpdatePass> MaybeParsePassword(std::string_view command,
+                                                     bool hashed = false) const;
+
+  std::optional<bool> MaybeParseStatus(std::string_view command) const;
+
+  using OptCat = std::optional<uint32_t>;
+  std::pair<OptCat, bool> MaybeParseAclCategory(std::string_view command) const;
+
+  using OptCommand = std::optional<std::pair<size_t, uint64_t>>;
+  std::pair<OptCommand, bool> MaybeParseAclCommand(std::string_view command,
+                                                   const CommandRegistry& registry) const;
+
+  std::variant<User::UpdateRequest, facade::ErrorReply> ParseAclSetUser(
+      facade::ArgRange args, const CommandRegistry& registry, bool hashed = false,
+      bool has_all_keys = false) const;
+
+  using MaterializedContents = std::optional<std::vector<std::vector<std::string_view>>>;
+
+  // Helper function that transforms a serialized acl file into a data structure we can
+  // iterate and push to the registry
+  MaterializedContents MaterializeFileContents(std::vector<std::string>* usernames,
+                                               std::string_view file_contents) const;
+
+  struct ParseKeyResult {
+    std::string glob;
+    KeyOp op;
+    bool all_keys{false};
+    bool reset_keys{false};
+  };
+
+  std::optional<ParseKeyResult> MaybeParseAclKey(std::string_view command) const;
+
+  std::string AclKeysToString(const AclKeys& keys) const;
+
+  std::string PasswordsToString(const absl::flat_hash_set<std::string>& passwords, bool nopass,
+                                bool full_sha) const;
+
+  struct CategoryAndMetadata {
+    User::CategoryChange change;
+    User::ChangeMetadata metadata;
+  };
+
+  struct CommandAndMetadata {
+    User::CommandChange change;
+    User::ChangeMetadata metadata;
+  };
+
+  using MergeResult = std::vector<std::variant<CategoryAndMetadata, CommandAndMetadata>>;
+
+  // Merge Category and Command changes and sort them by global order seq_no
+  MergeResult MergeTables(const User::CategoryChanges& categories,
+                          const User::CommandChanges& commands) const;
+
+  void BuildIndexers(RevCommandsIndexStore families);
+
+  // Data members
 
   facade::Listener* main_listener_{nullptr};
   UserRegistry* registry_;
   CommandRegistry* cmd_registry_;
   util::ProactorPool* pool_;
+
+  // Indexes
+
+  // See definitions for NONE and ALL in facade/acl_commands_def.h
+  const CategoryIndexTable cat_table_{{"KEYSPACE", KEYSPACE},
+                                      {"READ", READ},
+                                      {"WRITE", WRITE},
+                                      {"SET", SET},
+                                      {"SORTEDSET", SORTEDSET},
+                                      {"LIST", LIST},
+                                      {"HASH", HASH},
+                                      {"STRING", STRING},
+                                      {"BITMAP", BITMAP},
+                                      {"HYPERLOG", HYPERLOGLOG},
+                                      {"GEO", GEO},
+                                      {"STREAM", STREAM},
+                                      {"PUBSUB", PUBSUB},
+                                      {"ADMIN", ADMIN},
+                                      {"FAST", FAST},
+                                      {"SLOW", SLOW},
+                                      {"BLOCKING", BLOCKING},
+                                      {"DANGEROUS", DANGEROUS},
+                                      {"CONNECTION", CONNECTION},
+                                      {"TRANSACTION", TRANSACTION},
+                                      {"SCRIPTING", SCRIPTING},
+                                      {"BLOOM", BLOOM},
+                                      {"FT_SEARCH", FT_SEARCH},
+                                      {"THROTTLE", THROTTLE},
+                                      {"JSON", JSON},
+                                      {"ALL", ALL},
+                                      {"NONE", NONE}};
+
+  // bit 0 at index 0
+  // bit 1 at index 1
+  // bit n at index n
+  const ReverseCategoryIndexTable reverse_cat_table_{
+      "KEYSPACE",  "READ",      "WRITE",     "SET",       "SORTEDSET",  "LIST",        "HASH",
+      "STRING",    "BITMAP",    "HYPERLOG",  "GEO",       "STREAM",     "PUBSUB",      "ADMIN",
+      "FAST",      "SLOW",      "BLOCKING",  "DANGEROUS", "CONNECTION", "TRANSACTION", "SCRIPTING",
+      "_RESERVED", "_RESERVED", "_RESERVED", "_RESERVED", "_RESERVED",  "_RESERVED",   "_RESERVED",
+      "BLOOM",     "FT_SEARCH", "THROTTLE",  "JSON"};
+
+  // We need this to act as a const member, since the initialization of const data members
+  // must be done on the constructor. However, these are initialized a little later, when
+  // we Register the commands
+  const CategoryToIdxStore& CategoryToIdx(CategoryToIdxStore store = {}) const {
+    static CategoryToIdxStore cat_idx = std::move(store);
+    return cat_idx;
+  }
+
+  const RevCommandsIndexStore& CommandsRevIndexer(RevCommandsIndexStore store = {}) const {
+    static RevCommandsIndexStore rev_index_store = std::move(store);
+    return rev_index_store;
+  }
+
+  const CategoryToCommandsIndexStore& CategoryToCommandsIndex(
+      CategoryToCommandsIndexStore store = {}) const {
+    static CategoryToCommandsIndexStore index = std::move(store);
+    return index;
+  }
+
+  // Only for testing interface
+ public:
+  // Helper accessors for tests. Do not use them directly.
+  const ReverseCategoryIndexTable& GetRevTable() const {
+    return reverse_cat_table_;
+  }
+
+  // We could make CommandsRevIndexer public, but I want this to be
+  // clear that this is for TESTING so do not use this in the codebase
+  const RevCommandsIndexStore& GetCommandsRevIndexer() const {
+    return CommandsRevIndexer();
+  }
 };
 
 }  // namespace acl
