+++ src/server/acl/user_registry.cc
@@ -21,7 +21,7 @@ namespace dfly::acl {
 void UserRegistry::MaybeAddAndUpdate(std::string_view username, User::UpdateRequest req) {
   std::unique_lock<fb2::SharedMutex> lock(mu_);
   auto& user = registry_[username];
-  user.Update(std::move(req));
+  user.Update(std::move(req), *cat_to_id_table_, *reverse_cat_table_, *cat_to_commands_table_);
 }
 
 bool UserRegistry::RemoveUser(std::string_view username) {
@@ -79,11 +79,16 @@ User::UpdateRequest UserRegistry::DefaultUserUpdateRequest() const {
   return {std::move(pass), true, false, {std::move(acl)}, {std::move(key)}};
 }
 
-void UserRegistry::Init() {
+void UserRegistry::Init(const CategoryToIdxStore* cat_to_id_table,
+                        const ReverseCategoryIndexTable* reverse_cat_table,
+                        const CategoryToCommandsIndexStore* cat_to_commands_table) {
   // if there exists an acl file to load from, requirepass
   // will not overwrite the default's user password loaded from
   // that file. Loading the default's user password from a file
   // has higher priority than the deprecated flag
+  cat_to_id_table_ = cat_to_id_table;
+  reverse_cat_table_ = reverse_cat_table;
+  cat_to_commands_table_ = cat_to_commands_table;
   auto default_user = DefaultUserUpdateRequest();
   auto maybe_password = absl::GetFlag(FLAGS_requirepass);
   if (!maybe_password.empty()) {
